requirements.txt
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.8.2
pydantic-settings>=2.4.0
tomli>=2.0.1; python_version < '3.11'
google-generativeai>=0.7.2
backoff>=2.2.1
python-slugify>=8.0.4
pandas>=2.2.2
